Planning Your Website.
The goals of web design.
To understand how to create well designed websites, it is worth understanding thef undamentals of design. Good design creates experiences that makes people’s lives easier, and is also aesthetically pleasing. Here are some examples: a well-designed store makes it easy for customers to fi nd products they are looking for, or maybe even products they never knew they wanted. Drivers can see well-designed road signs from a distance and the message of the sign can be immediately understood, even at a high rate of speed while driving. A welldesigned chair is comfortable to sit in and also fi ts in with the style of the room.